SshMachineLocationTest: make non-integration
* Uses RecordingSshTool
* Moves integration tests into SshMachineLocationIntegrationTest
* Changes SshMachineLocationIntegrationTest to extends 
  SshMachineLocationTest.
